Service improvements, pdf-read
