--- templates/convert-pdf.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Convert PDF</title>
-</head>
-<body>
-    <h2>PDF to Convert:</h2>
-    {% for image in images %}
-        <img src="data-image/jpeg;base64,{{ image }}" alt="PDF image">
-    {% endfor %}
-    <form action="/pdf/convert-pdf" method="post">
-        <input type="hidden" name="file_path" value="{{ cache_key }}">
-        <input type="submit" value="Convert PDF">
-    </form>
-</body>
-</html>
