--- templates/upload-pdf.html
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>PDF Preview</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='pdf.js/build/pdf.worker.css') }}"> 
-    </head>
-<body>
-    <h1>PDF Preview</h1>
-    <form id="upload-form" method="post" action="/upload-pdf" enctype="multipart/form-data">
-        <input type="file" name="file" required>
-        <input type="submit" value="Upload PDF">
-    </form>
-
-    <div id="pdf-viewer"></div> 
-
-    <script src="{{ url_for('static', filename='pdf.js/build/pdf.js') }}"></script> 
-    <script src="{{ url_for('static', filename='pdf.js/build/pdf.worker.js') }}"></script> 
-    <script>
-        let currentPDF = null; // Store currently displayed PDF
-
-        function displayPDF(pdfURL) {
-            pdfjsLib.GlobalWorkerOptions.workerSrc = "{{ url_for('static', filename='pdf.js/build/pdf.worker.js') }}";
-
-            const loadingTask = pdfjsLib.getDocument(pdfURL);
-            loadingTask.promise.then(function(pdf) {
-                currentPDF = pdf;
-                renderPage(1); // Render first page initially
-            }, function (error) {
-                console.error("Error loading PDF:", error);
-            });
-        }
-
-        function renderPage(pageNumber) {
-            currentPDF.getPage(pageNumber).then(function(page) {
-                const scale = 1.5; // Adjust scale as needed
-                const viewport = page.getViewport({ scale: scale });
-                const canvas = document.createElement('canvas');
-                const context = canvas.getContext('2d');
-                canvas.height = viewport.height;
-                canvas.width = viewport.width;
-
-                const renderTask = page.render({ canvasContext: context, viewport: viewport });
-                renderTask.promise.then(function () {
-                    $('#pdf-viewer').empty().append(canvas);
-                });
-            });
-        }
-
-        $('#upload-form').submit(function(e) {
-            e.preventDefault();
-            // ... (Your AJAX upload logic) ...
-            success: function(response) {
-                // Assuming your backend returns the temporary file URL
-                const pdfURL = response.file_url; 
-                displayPDF(pdfURL);
-            }
-        });
-    </script>
-</body>
-</html>
