--- templates/upload2-pdf.html
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>PDF Preview</title>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
-    <!-- <link rel="stylesheet" href="{{ url_for('static', filename='styles/style.css')}}"> -->
-</head>
-<body>
-    <h1>PDF Preview and Conversion</h1>
-    <form id="upload-form" method="post" action="/upload-pdf" enctype="multipart/form-data">
-        <input type="file" name="file" required>
-        <input type="submit" value="Upload PDF">
-    </form>
-
-    <div id="preview-area" style="display: none;">
-        <h2>Preview</h2>
-        <div id="image-previews"></div>
-        <button id="convert-button" disabled>Convert to SVG</button>
-    </div>
-
-    <script>
-        $('#upload-form').submit(function(e) {
-            e.preventDefault();
-            const formData = new FormData(this);
-
-            $.ajax({
-                url: '/upload-pdf',
-                type: 'POST',
-                data: formData,
-                cache: false,
-                contentType: false,
-                processData: false,
-                success: function(response) { 
-                    if(response.images) {
-                        $('#image-previews').empty(); 
-                        response.images.forEach(image => {
-                            const imgElement = $('<img>').attr('src', 'data:image/jpeg;base64,' + image);
-                            $('#image-previews').append(imgElement);
-                        });
-                        $('#convert-button').attr('disabled', false);  // Enable convert button
-                        $('#preview-area').show();
-                    } else {
-                        // Handle errors if needed
-                        alert(response.error || "An error occurred");
-                    }
-                },
-                error: function() {
-                    alert("Blad przy wczytywaniu pliku lub generowaniu podglądu");
-                }
-            });
-        });
-        
-        $('#convert-button').click(function() {
-            const cacheKey = '{{ cache_key }}';
-            window.location.href=`/convert-pdf/${cacheKey}`;
-        });
-    </script>
-</body>
-</html>
